+++ packages/apply-release-plan/src/index.ts
@@ -71,7 +71,8 @@ export default async function applyReleasePlan(
   releasePlan: ReleasePlan,
   packages: Packages,
   config: Config = defaultConfig,
-  snapshot?: string | boolean
+  snapshot?: string | boolean,
+  contextDir = __dirname
 ) {
   let cwd = packages.root.dir;
 
@@ -100,7 +101,8 @@ export default async function applyReleasePlan(
     releasesWithPackage,
     changesets,
     config,
-    cwd
+    cwd,
+    contextDir
   );
 
   if (releasePlan.preState !== undefined && snapshot === undefined) {
@@ -136,7 +138,8 @@ export default async function applyReleasePlan(
     });
   });
 
-  let prettierInstance = getPrettierInstance(cwd);
+  let prettierInstance =
+    config.prettier !== false ? getPrettierInstance(cwd) : undefined;
 
   for (let release of finalisedRelease) {
     let { changelog, packageJson, dir, name } = release;
@@ -205,7 +208,8 @@ async function getNewChangelogEntry(
   releasesWithPackage: ModCompWithPackage[],
   changesets: NewChangeset[],
   config: Config,
-  cwd: string
+  cwd: string,
+  contextDir: string
 ) {
   if (!config.changelog) {
     return Promise.resolve(
@@ -223,7 +227,13 @@ async function getNewChangelogEntry(
 
   const changelogOpts = config.changelog[1];
   let changesetPath = path.join(cwd, ".changeset");
-  let changelogPath = resolveFrom(changesetPath, config.changelog[0]);
+  let changelogPath;
+
+  try {
+    changelogPath = resolveFrom(changesetPath, config.changelog[0]);
+  } catch {
+    changelogPath = resolveFrom(contextDir, config.changelog[0]);
+  }
 
   let possibleChangelogFunc = require(changelogPath);
   if (possibleChangelogFunc.default) {
@@ -283,7 +293,7 @@ async function updateChangelog(
   changelogPath: string,
   changelog: string,
   name: string,
-  prettierInstance: typeof prettier
+  prettierInstance: typeof prettier | undefined
 ) {
   let templateString = `\n\n${changelog.trim()}\n`;
   let fileData;
@@ -337,15 +347,17 @@ async function updatePackageJson(
 async function writeFormattedMarkdownFile(
   filePath: string,
   content: string,
-  prettierInstance: typeof prettier
+  prettierInstance: typeof prettier | undefined
 ) {
   await fs.writeFile(
     filePath,
-    // Prettier v3 returns a promise
-    await prettierInstance.format(content, {
-      ...(await prettierInstance.resolveConfig(filePath)),
-      filepath: filePath,
-      parser: "markdown",
-    })
+    prettierInstance
+      ? // Prettier v3 returns a promise
+        await prettierInstance.format(content, {
+          ...(await prettierInstance.resolveConfig(filePath)),
+          filepath: filePath,
+          parser: "markdown",
+        })
+      : content
   );
 }