+++ packages/cli/src/commands/version/index.ts
@@ -7,8 +7,6 @@ import applyReleasePlan from "@changesets/apply-release-plan";
 import readChangesets from "@changesets/read";
 import assembleReleasePlan from "@changesets/assemble-release-plan";
 import { getPackages } from "@manypkg/get-packages";
-
-import { removeEmptyFolders } from "../../utils/v1-legacy/removeFolders.ts";
 import { readPreState } from "@changesets/pre";
 import { ExitError } from "@changesets/errors";
 import { getCommitFunctions } from "../../commit/getCommitFunctions.ts";
@@ -37,7 +35,6 @@ export default async function version(
   const [changesets, preState] = await Promise.all([
     readChangesets(cwd),
     readPreState(cwd),
-    removeEmptyFolders(path.resolve(cwd, ".changeset")),
   ]);
 
   if (preState?.mode === "pre") {
@@ -85,7 +82,8 @@ export default async function version(
     releasePlan,
     packages,
     releaseConfig,
-    options.snapshot
+    options.snapshot,
+    __dirname
   );
 
   const [{ getVersionMessage }, commitOpts] = getCommitFunctions(