+++ packages/config/src/index.ts
@@ -3,7 +3,7 @@ import path from "path";
 import micromatch from "micromatch";
 import { ValidationError } from "@changesets/errors";
 import { warn } from "@changesets/logger";
-import type { Packages } from "@manypkg/get-packages";
+import { type Packages, getPackages } from "@manypkg/get-packages";
 import type {
   Config,
   WrittenConfig,
@@ -91,7 +91,8 @@ function isArray<T>(
   return Array.isArray(arg);
 }
 
-export let read = async (cwd: string, packages: Packages) => {
+export let read = async (cwd: string, packages?: Packages) => {
+  packages ??= await getPackages(cwd);
   let json = JSON.parse(
     await fs.readFile(path.join(cwd, ".changeset", "config.json"), "utf8")
   );
@@ -332,6 +333,16 @@ export let parse = (json: WrittenConfig, packages: Packages): Config => {
     }
   }
 
+  if (json.prettier !== undefined) {
+    messages.push(
+      `The \`prettier\` option is set as ${JSON.stringify(
+        json.prettier,
+        null,
+        2
+      )} when the only valid values are undefined or a boolean`
+    );
+  }
+
   const { snapshot } = json;
 
   if (snapshot !== undefined) {
@@ -480,6 +491,8 @@ export let parse = (json: WrittenConfig, packages: Packages): Config => {
         json.___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH
           ?.updateInternalDependents ?? "out-of-range",
     },
+
+    prettier: typeof json.prettier === "boolean" ? json.prettier : true,
 
     // TODO consider enabling this by default in the next major version
     privatePackages: