+++ packages/write/src/index.ts
@@ -20,7 +20,8 @@ function getPrettierInstance(cwd: string): typeof prettier {
 
 async function writeChangeset(
   changeset: Changeset,
-  cwd: string
+  cwd: string,
+  options?: { prettier?: boolean }
 ): Promise<string> {
   const { summary, releases } = changeset;
 
@@ -33,7 +34,8 @@ async function writeChangeset(
     capitalize: false,
   });
 
-  const prettierInstance = getPrettierInstance(cwd);
+  const prettierInstance =
+    options?.prettier !== false ? getPrettierInstance(cwd) : undefined;
   const newChangesetPath = path.resolve(changesetBase, `${changesetID}.md`);
 
   // NOTE: The quotation marks in here are really important even though they are
@@ -49,11 +51,13 @@ ${summary}
   await fs.mkdir(path.dirname(newChangesetPath), { recursive: true });
   await fs.writeFile(
     newChangesetPath,
-    // Prettier v3 returns a promise
-    await prettierInstance.format(changesetContents, {
-      ...(await prettierInstance.resolveConfig(newChangesetPath)),
-      parser: "markdown",
-    })
+    prettierInstance
+      ? // Prettier v3 returns a promise
+        await prettierInstance.format(changesetContents, {
+          ...(await prettierInstance.resolveConfig(newChangesetPath)),
+          parser: "markdown",
+        })
+      : changesetContents
   );
 
   return changesetID;