+++ packages/should-skip-package/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @changesets/should-skip-package
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`84a4a1b`](https://github.com/changesets/changesets/commit/84a4a1b1d399bfd0a58677b0182b9c053194febf)]:
+  - @changesets/types@6.1.0
+
 ## 0.1.1
 
 ### Patch Changes