+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @changesets/types
 
+## 6.1.0
+
+### Minor Changes
+
+- [#1453](https://github.com/changesets/changesets/pull/1453) [`84a4a1b`](https://github.com/changesets/changesets/commit/84a4a1b1d399bfd0a58677b0182b9c053194febf) Thanks [@bennypowers](https://github.com/bennypowers)! - Added a new config option to opt-out from formatting with Prettier using `prettier: false`.
+
 ## 6.0.0
 
 ### Major Changes